# projects/nikway/Nikway Academy Sep 26/continuous-mission/publish-ready-sample/generated-openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_evidence_api_v1_assignments__assignment_id__evidence_upload_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_evidence_api_v1_assignments__assignment_id__evidence_upload_post",
        "type": "object"
      },
      "EvidenceInput": {
        "properties": {
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "evidence_type": {
            "title": "Evidence Type",
            "type": "string"
          },
          "object_storage_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Storage Ref"
          },
          "step_id": {
            "title": "Step Id",
            "type": "string"
          }
        },
        "required": [
          "step_id",
          "evidence_type"
        ],
        "title": "EvidenceInput",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "NIKWAY V1 Pilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assessment-results/{assessment_id}/progression": {
      "post": {
        "operationId": "decide_progression_api_v1_assessment_results__assessment_id__progression_post",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": true,
            "schema": {
              "title": "X-Organization-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Decide Progression Api V1 Assessment Results  Assessment Id  Progression Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Progression"
      }
    },
    "/api/v1/assignments/{assignment_id}": {
      "get": {
        "operationId": "get_assignment_api_v1_assignments__assignment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": true,
            "schema": {
              "title": "X-Organization-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Assignment Api V1 Assignments  Assignment Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Assignment"
      }
    },
    "/api/v1/assignments/{assignment_id}/evidence": {
      "post": {
        "operationId": "submit_evidence_api_v1_assignments__assignment_id__evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": true,
            "schema": {
              "title": "X-Organization-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Evidence Api V1 Assignments  Assignment Id  Evidence Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Evidence"
      }
    },
    "/api/v1/assignments/{assignment_id}/evidence/upload": {
      "post": {
        "operationId": "upload_evidence_api_v1_assignments__assignment_id__evidence_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "step_id",
            "required": true,
            "schema": {
              "title": "Step Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": true,
            "schema": {
              "title": "X-Organization-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_evidence_api_v1_assignments__assignment_id__evidence_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Evidence Api V1 Assignments  Assignment Id  Evidence Upload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Evidence"
      }
    },
...
    "/api/v1/masters/{master_id}": {
      "get": {
        "operationId": "get_master_api_v1_masters__master_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "master_id",
            "required": true,
            "schema": {
              "title": "Master Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-Id",
            "required": true,
            "schema": {
              "title": "X-Organization-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Master Api V1 Masters  Master Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Master"
      }
    },
...
  }
}